Updated openstack/openstack Project: openstack/designate  8ca54f40109d15a2c7258d185d362ea38a2b8112
Ensure test_find_zone_imports waits for all imports

test_find_zone_imports was only waiting for a single zone
import to complete, rather than both of the imports created
during the tests to complete. This resulted in occasional
failures in the gate if the second import finished before
the first.

Change-Id: Iad2b518b3e4bccded305de2ee3e7fdbbf82095f8
Closes-Bug: 1470854