Updated openstack/openstack Project: openstack-infra/devstack-gate  0838a01c587b45b9f7c6febf06ca7ae81b3e7118
Treat bug/ and feature/ branches the same way

Add support for bug/ branches for projects that use intermediate
releases to support the case where a change has landed in master and a
bug fix is needed, but the branch to be created should not be seen as a
long term stable release series. For example, if a breaking change is
made in a library in anticipation of a major series version update, but
then a critical bug is found in the previous intermediate release. The
fix could be back-ported and released into the master stream without the
breaking change.

Change-Id: Ic790600c4c058817e32422f7c7d1be7fdbf98994