Updated openstack/openstack Project: openstack/tempest  f0ccd94275e03ae32a8bf9cc451ccde04f96aa26
Apply a naming rule of GET to compute clients(t*)

[GET /resources] methods should be "list_<resource name>s"
or "show_<resource name>", so this patch applies the rule
to compute clients which names are "t*".

In addition, this patch removes unnecessary str() because
all callers specify string values and we need to pass py3
tests when migrating them into tempest-lib.

Partially implements blueprint consistent-service-method-names

Change-Id: I9d7787d50f616e74ccbc06388813fc7fcb994dea