Updated openstack/openstack Project: openstack/tempest  3ea405260a2c34a2847c4d35772415a37a9c2076
Apply a naming rule of GET to compute clients(v*)

[GET /resources] methods should be "list_<resource name>s"
or "show_<resource name>", so this patch applies the rule
to compute clients which names are "v*".

In addition, this patch removes unnecessary str() because
all callers specify string values and we need to pass py3
tests when migrating them into tempest-lib.

Partially implements blueprint consistent-service-method-names

Change-Id: I17b62791b287bb7150603a029049278756a8616d