Updated openstack/openstack Project: openstack/tuskar  64926ca8e04b686896aecf9afe910ae52d64ea93
Added --all flag to tuskar-delete-roles

This is mostly for development purposes, but when working on the TripleO
Heat Templates, it's common to flush out all of the existing roles from
Tuskar and load in the changed ones. It was tedious to do when
specifying all of the UUIDs, so this is provided as a shortcut.

Change-Id: Ib8063eb7aa737a71b208a142318ea62c02b039a7