Updated openstack/openstack Project: openstack-infra/puppet-jenkins  e96caaa829c0dc41edb00997c0470651150d1a89
Pin specific git-review version instead of latest

The puppet pip provider infers ensure=>latest to mean use pip
install --upgrade. This has the unfortunate side effect of upgrading
any install_requires of the thing being installed even if the
requirement is already met (for example, by existing distro packages
on the system). This addresses a regression where git-review was
causing the distro-installed requests 1.1 to be upgraded even though
git-review claims to work with that version.

Change-Id: I64dd265543ef401d057fca7183363b9f31cdf523