Updated openstack/openstack Project: openstack/diskimage-builder  e33c3d96ad962d355caa01dc20a8d2bdade6830d
Document only exectuables in phase dirs policy

Having data files in the phase subdirs is an easy source of confusion
in reviews (especially when the data file is a script) and theres really
no reason to be putting data files there at this point. Lets make a
convention out of not doing this.

Change-Id: I99571a2a49e14e8c709af20f6d13d662ac745eb4