Updated openstack/openstack Project: openstack/nova  68a614c4dff4e9388591be345d38d51245a2d864
Convert RT compute_node to be a ComputeNode object

This patch converts the ResourceTracker compute_node property
to be a ComputeNode object. A number of fields automatically
take care of mapping their values to a db format, so some of the
code creating json strings goes away with this change.

The scheduler client report code is simplified by the change
to use a ComputeNode object.

Note that this change naturally required modification to a
number of tests in test_tracker, test_resource_tracker and
test_client to cater for objects instead of dicts. Some of these
tests were using incorrect values or arbitrary key names that do
not exist as ComputeNode fields, so they had to be corrected to
conform to the type checking of the ComputeNode object.

part of blueprint make-resource-tracker-use-objects

Change-Id: I2279f01ad55083c31c663242a2a60a48191e88c3