Updated openstack/openstack Project: openstack/nova  bc3806e6dd0c7deba186c7a528edab4c9c76bb22
Remove compat_instance()

This was added to support older RPC calls where instance_properties had some
flavor stuff embedded in system_metadata. That is not (AFAIK) still required
after the move to compute RPC v4.0 and the dropping of v3.0 support.

This is required to drop flavors-in-sysmeta compatibility across the board.

Note that this attempts to keep the scheduler's request_spec contract for
RPC by stashing old-style flavor info into the system_metadata of the
instance_properties if we have a full flavor. We don't think it's ever
actually used (at least by in-tree code), but this change attempts to be
well-behaved in this regard.

Change-Id: I246b23ef48227df7e775018c116afe656a0d1633