Updated openstack/openstack Project: openstack/heat  10225ce417aa6172d18afb63ab748ae340a23c48
In-tree grenade support for Heat

The following implements in tree support for grenade upgrade testing
for heat. It includes the expected directory structure to be used for
the external plugin.

For now, we create a simple OS::Heat::RandomString resource
and check whether the stack survives the upgrade.
We should add better tests for upgrades in the future.

Co-Authored-By: Steven Hardy <shardy@redhat.com>
Co-Authored-By: Sergey Kraynev <skraynev@mirantis.com>
Co-Authored-By: Emilien Macchi <emilien.macchi@enovance.com>

partial blueprint upgrade-tests

Change-Id: I0847004a8ac023a113cabaee46896abc823a01da