Updated openstack/openstack Project: openstack-infra/system-config  ef0529198cacea29392eea8c2ed7e0f9c63f7d57
Use the cache-bindep element on dib nodes

We intend to use bindep to tell us what distro packages should be
installed at job run-time, so need to similarly identify those in
such a way that they can be cached on our worker images.

Install the cache-bindep element on all diskimage-builder based
nodepool images, so that we will have bindep available to jobs which
need it along with a cache of the distro packages they might
require.

Note that similar tooling is not applied to snapshot nodes, since we
only intend to run jobs which use bindep on diskimage-builder
created nodes.

Change-Id: I22b39f5117bb78d387414668c63df135cbc53243
Depends-On: I0c408fee35dfe9b4f700c51565bfaae0a3d03beb