Updated openstack/openstack Project: openstack-infra/zuul  1c42a2ab4704ecce112aa469dc7e1d13aa80d140
Add uuid to the launch log record

After a particular change gets scheduled, launch records for each job
that will run for that change are logged.  Subsequent build records will
then be logged to capture state about the running job.  By including the
uuid in the launch log record, it will become possible to correlate the
launch record for a job with the build records for that job for a given
change using a log parser.

Change-Id: I24e1ad43a5c7008a7e0587f5df1d5f332a8e4ac3