Updated openstack/openstack Project: openstack/nova  954986395a6dc7dd6a559530c7f55de79c8d31a7
Handle unexpected clear events call

Currently we don't check if we have access to events at the time of
calling clear_events for an instance. This could result in operations
failing (like deleting an instance) for the user.

Add a check and a log message that this shouldn't happen (but it does).

Change-Id: I29852662cbfe38f37f9d5154147e6b17b7b8b454
Closes-Bug: 1454149