Updated openstack/openstack Project: openstack/sahara-specs  3aa093dbc3814d3aabfa2eb534cab98c13eab3db
Use Templates for Scenario Tests Configuration

Users that want to run the scenario tests available in the Sahara repository
need to modify the provided YAML files, which are really template files even
if not marked as such. The usability of this process could be improved by
extending the test runner to read the templates and perform the substitution
of the required variables instead.

Change-Id: I5ea728d14a1727c8c6f5c07f320e20fb9b1cb218
Describes: blueprint scenario-test-config-template