Updated openstack/openstack Project: openstack-infra/puppet-accessbot  4d84c482111c3732cb21d0f89e2245bcd3cee527
Update template to use instance variables

Using methods to access in-scope variables from puppet templates has
been deprecated since puppet 3.2.0[1] and causes warnings to be emitted
in logs. In puppet 4 this functionality was removed entirely. This
patch updates the accessbot.confg template to use instance variables as
recommended in the puppet documentation[2].

[1] https://projects.puppetlabs.com/issues/19058
[2] https://docs.puppetlabs.com/guides/templating.html#referencing-variables

Change-Id: I7af52e32322333879eef946443620e09a726dd57