Updated openstack/openstack Project: openstack-infra/project-config  43e8e90cbf717cdd47452d721e422704564341f1
Don't fail to build cache on missing bindep lists

While git show is being used to build up a list of candidate
packages to cache, it can cause the cache hook to abort if it
returns nonzero because of a nonexistent bindep list file. Guard
that with a fallback to true so that the loop can continue
normally.

Change-Id: Ic3b773ad2e8efa2c450d37a0115355c0e95da853