Updated openstack/openstack Project: openstack-infra/project-config  6cedb7ea12d5df50e099d7b50aec0baa1ed2bd43
Create automated job for sample configuration files

The keystone team updates the sample configuration file in it's
own patch, so to not stomp on other patches.

We've had to update the configuration files manually, but we
should create an automated job. Since it's just running
tox -e genconfig and checking diff output.

Change-Id: I0e294b857d9bbf73e7f9c0d0bdee61648fb999bd