Updated openstack/openstack Project: openstack/neutron  e9c4b6d974694e89d4a3bdefd33aea7445387490
Fall back on empty path if prefix is missing

A missing entry causes a KeyError that leads the server to blow up during
startup. We can fallback on an empty path (like some services do), in case
the prefix is not specified.

Furthermore, we can be declarative with this property, the same way we are
with properties like aliases, bulk support, etc.

Change-Id: I58a9b90a39d434f4808264aeb6f9ee5aceff7fbd