Updated openstack/openstack Project: openstack/keystone  33168d6582641219d258f5a637aafd10e6bce04b
Add test case for deleting endpoint with space in url

In the bug report, creating an endpoint with "tenant_i d" in any of
its URLs (internal, public, or admin) allegedly prevented the
endpoint from being deleted.

This test case adds a space to all 3 URLs of an endpoint. It then
deletes this endpoint and attempts to get it. If the get returns
error 404, then it works.

Change-Id: I0cad2410a150fb6967fc5048ebbf00b9b08efcc7
Closes-Bug: #1098564