Updated openstack/openstack Project: openstack/tempest  9821a27d68212f7861c51f15818a93a24ce9ee57
Get rid of duplicated creation of security groups in test_volume_boot_pattern

It turns out that a security group in test test_volume_boot_pattern
is created twice. The first security group is created in test itself,
the second security group is created in the method _boot_instance_from_
volume(). This patch fixes it. Now the security group is created once.

Change-Id: I1c10e80161880c7ad04f7c0f48d51cc03b6cbada