Updated openstack/openstack Project: openstack/python-neutronclient  043656c8f8945b5c236bbd30b460f94e05a5a59d
Determine ip version during subnet create.

Currently, the default IP version for a subnet-create is IPv4.
This behavior is wrong when we use a IPv6 subnetpool and
do not specify the ip version explicitly during the subnet
create operation.

This fix proposes to make the neutron client code for subnet
creation become aware of the subnetpool version. This ensures
that the proper IP version is sent down to the neutron server
if the subnet create is requested with a subnetpool.

Co-Authored-By: Akihiro Motoki <amotoki@gmail.com>

Change-Id: I13da0f204e8ce335a2082c7d829bee28ac567c3f
Closes-bug: 1444146