Updated openstack/openstack Project: openstack/nova  d03ef34b0b1ed96a2f2bea1f5f01f09436c55125
API: make sure a blank volume with no size is rejected

Currently - we would happily accept the request, but fail to create the
block device mapping record for such a volume. The result was that the
boot request for such an instance would be accepted, but would not
create a volume.

Also fix the _volume_size API method that was 0ing out blank volume's
size by mistake.

Change-Id: I1d243aed78da7c026e55ea145397b070d619a567
Closes-bug: #1461638