Updated openstack/openstack Project: openstack/neutron-specs  6b3edd7a14175a9f2bc507ed183f63971f6bb7d4
Don't put subnet pools in a scope by default

The problem is that subnet pools are already out there and people are
likely to use them before address scopes are available.  To introduce
address scopes and immediately apply the routing restrictions to
existing subnet pools in Liberty could come as quite a shock to
existing users.

I'm putting this up as a change to the BP to get opinions and see if
they align.

Change-Id: I2fbdf31a818c8d96af78a47778aea4a24c24cbf3