Updated openstack/openstack Project: openstack/tempest  e9dee49a90c71303eae8a61637b5a6375866ed95
Re-factor neutron client for 'show' methods

Currently neutron client is written to auto generate
methods on fly. This is very difficult to debug and
make code very difficult to read.

As neutron client will be moved to lib, it will be better to
have its interfaces more readable and consistent with other
service clients.

This commit re-factor it for 'show_*' methods.

Change-Id: I953ef3fdbae7b9d630b1de2a324142ad00d29b57