Updated openstack/openstack Project: openstack/tempest  17a8e408e87cb7ffa08f72a4f20cc1fcd5c5fb26
Re-factor neutron client for 'delete' methods

Currently neutron client is written to auto generate
methods on fly. This is very difficult to debug and
make code very difficult to read.

As neutron client will be moved to lib, it will be better to
have its interfaces more readable and consistent with other
service clients.

This commit re-factor it for 'delete_*' methods.

Change-Id: I8b37555a250e5988553c3cdcfc4851487f27ea38