Updated openstack/openstack Project: openstack/glance  987f874ac4ed72a88e0e76ecad796aaa88125e96
Change default digest_algorithm value to sha256

In I9236cc85f4e9881ac1aa35d69bc6761a59c1b6c8 it was promised that the
default for digest_algorithm would change from sha1 to sha256. The sole
purpose of this commit is to upgrade the default to sha256 as promised.

DocImpact
SecurityImpact
Change-Id: I12be91db7b86e335a84c9ebed86dac3ba09051cb