Updated openstack/openstack Project: openstack/python-neutronclient  31f8f23fb5d52a2d674b9810f413c9d2afb5eacf
Avoid overwriting parsed_args

I just notice this in the context of this review [1].  It seemed like an odd
pattern to me but I wasn't sure if it should be folded in to that patch because
of the reach of the change.  So, I thought I'd post it as a clean-up.

[1] https://review.openstack.org/#/c/194078/13/neutronclient/neutron/v2_0/subnet.py

Change-Id: Iaf53f595e6d50fc18a33258439b6ae5e476dfca9