Updated openstack/openstack Project: openstack/oslo.concurrency  7243f588103c3728172dde23ebebae218fc51046
processutils: ensure on_completion callback is always called

If the subprocess.Popen.communicate method raises an exception,
the on_completion callback is never invoked. If a caller is
trying to use on_execute + on_completion to track lifecycle
of a process this creates a problem, as they cannot reliably
detect completion.

Change-Id: I22b2d7bde8797276f7670bc289d915dab5122481
Closes-bug: #1470868