Updated openstack/openstack Project: openstack/keystonemiddleware  4e7c4ed250a2c586f817c84fa47f5ed975a4b2c8
Don't allow webob to set a default content type

Webob sets the Content-Type of responses to 'text/html' if nothing else
is set. There are a number of bugs out for webob with this but it seems
unlikely it will be fixed upstream.

We don't want auth_token middleware automatically setting content-types
on application responses so create our own response object with an empty
default_content_type so it's not set.

Change-Id: I947f384e170a8d2685b71e1a6611c6241e827196
Closes-Bug: #1466499