Updated openstack/openstack Project: openstack-dev/pbr  c5cc43a5beacb7c1d3e147f3654fdea207e8340a
This documents the use of comments that work between 2.6, 2.7 and 3.4

Comments are not directly handled by PBR, this documents a reasonable working
set of comments that can be used in setup.cfg.

Comments are parsed out over three locations, configparser, which varies rather
significantly between two and three, pkg_resources, and lastly when pip runs.

Change-Id: Iedebc5baca4e21081708971540d6eab89e043f05