Updated openstack/openstack Project: openstack/heat  887be1716a6bc55a5831acc7d28979bc5a9b4b41
Keep the path when re-registering the template resource

Since Ib0c191db2dccac6d467a9961062424bc1f924c24 we re-register a
template resource in case the global environment is changed. This
is fine but the path to the resource is not maintained. This patch
fixes that by passing the path through to the registration function.

Change-Id: I7a8ec66d0f12a12bb4e44d2eb7efd182d758fea1