Updated openstack/openstack Project: openstack/qa-specs  f23cdfaa763f4c2be5b2f850d5040d47af7fe8d9
Move devstack-external-plugins spec into implemented

This commit moves the devstack-external-plugins spec into the
implemented dir. This bp was closed some time ago but the spec
approval fell through the cracks and was only merged recently.
Additionally the recent patch to do a mass cleanup of all the
specs missed this move to implemented. So let's take care of
it now.

As part of this the devstack specs line in the index is commented
out because this was the last open devstack spec. Without any open
specs in the devstack dir sphinx will warn on there being no files,
which we treat as a fatal error in ci.

Change-Id: Icf4d96ac46c5abdec5f7036d10d5e249d8471a73