Updated openstack/openstack Project: openstack/ironic  c6608e97af33f8d7f3fe2f7deeb78f52196f2cc7
Replace is_hostname_safe with a better check

is_hostname_safe was broken in several ways, which are documented in
bug #1468508. This replaces is_hostname_safe with a
is_valid_logical_name, which checks that logical names contain only
RFC 3986 unreserved characters, which is what was decided upon in the
Ironic meeting on 29 June 2015.

A new API microversion, 1.10, is introduced to handle this change.

Additionally, the performance of the old is_hostname_safe() is
improved in several ways.

Closes-Bug: #1468508
Closes-Bug: #1434376
Change-Id: I77441ae5a00e52b84051672ecab5785dd0b033a5