Updated openstack/openstack Project: openstack/barbican  8825b0378e6e4c62e5b7d1ed102be933e1b253e3
Fix JSON structure example for stored-key orders

The JSON dict that's used for the meta that's given to the server to
make a stored-key typed certificate order should be 'container_ref'
and not 'source_container_ref'. This confusion might have been
introduced since the client actually refers to that as the former
('source_container_ref'), even though the actual JSON that is sent to
the server is different.

Change-Id: Icb8d2052c9e8aee72ce48709b3ee40bcc188a85b