Updated openstack/openstack Project: openstack/horizon  a385be24f5286ef5e0edd00009d9dfefd1014cb7
corrected the wrong url in admin instance detail

In admin instance detail page, the url for attached volume
pointing to volume detail in project panel.

This behavior is wrong, it should be link to admin panel only.

This patch corrected the url to point to the admin panel volume detail.

Change-Id: I76419f1236304211602fa98f3885d8a2a05b6468
Closes-Bug: #1470821