Updated openstack/openstack Project: openstack/ironic  70e303741453c90cd5df36017797e35dbd9409d6
Image service should not be set in ImageCache constructor

It is possible to initialize ImageCache with specific image service
class that will be downloading images. When there was only one type
of image links supported - links from glance, it was OK, but now http
and file links are also supported, so if ImageCache will be
initialized with specific image service (e.g. glance), it will not be
able to download other supported types of images. This should not be
allowed.

image_service argument in common.images.fetch is removed for the same
reason. As a consequence of that, out-of-tree image services should
be added to common.image_service.get_image_service method so that
they can be used properly by common.images.fetch.

Closes-bug: #1468727

Change-Id: I428daa4ddf6f88ef222e8b26650d0ad461a67eec