Updated openstack/openstack Project: openstack-infra/elastic-recheck  a2e79fdb6d7c9f19c926bd1b3b403128de12d63a
Remove obsolete cinder LVM no space  querie

This querie shows up in a number of expected tests which
pass including Manilla and neutron-rally tests.

Regardles, looking through the queries and I'm not seeing
"dead" Cinder services, and don't recall actually having one
since we moved the LVM off loopbacks and made some other
changes to the encryption tests in Tempest.

Generic queries like this might be "easy" but the fact is that
looking at the results, in most cases this seems to be a
cascade affect from an earlier failure.  By logging everyhing
against a general issue like we are here we're actually masking
issues and they'll never get fixed until we remove the general
resultant failure and log failures/ER's for the root cause.

Change-Id: I936784890c9ddae990a26392bd3864913d90fba6