Updated openstack/openstack Project: openstack-dev/devstack  6d7b259094dcf0f8326b7eb14d239bdcdbacf458
Search for service by type

When you get or create service it first checks to see if an existing
service matching these parameters exists. The definition of existing is
having a service with the same name, however name is not a unique field.
Sahara for example creates two services, one with data-processing, one
with data_processing with the same sahara name.

Search for existing services by service type, not by service name.

Change-Id: I6148e2254aa3968039b0e7c178e7cabc53b6be68