Updated openstack/openstack Project: openstack-infra/release-tools  17919b75b5fd0055c1ee512767100b9e8f366a13
Close milestones for post-version releases

In the process of cleaning up the alpha version handling the call to
close the milestone was removed. The old call was to upload_milestone.py
with --nop to cause the upload to be skipped. Rather than restoring it
directly, create a new script that only closes the milestone and use
that.

Change-Id: Idf7bef35e290d2888800d33aba17ca744c90bfc9