Updated openstack/openstack Project: openstack-infra/release-tools  636c93c80e58ff3c8b2d2477bb7c6c803c89a8d7
Handle more errors when updating bugs

Sometimes the user running the script doesn't have access to a bug that
is marked private and gets a 401 error. Catch *all* errors trying to
update the bug status and treat them all as failures from which we can
continue and then clean up manually.

Change-Id: Ibc37d69259e841ba80bf4d42de5b0b45678fc54b