Updated openstack/openstack Project: openstack/cinder-specs  0982dec28683ebf8c3ec0c085a8b874f664ef178
Volume migration improvement for Liberty version

This spec proposes to make volume migration more stable, robust and
user friendly in terms of adding the following features.
 * A better volume status management
 * Migration progress indication
 * An efficient way of migration
 * Migration status and migration progress look up
 * Notification to Ceilometer
 * Tempest tests & CI support, and so on.

All the cases within this spec are targeted to the available volume.
There is going to be another spec to cover the issues regarding the
attached volume.

Related-BP: migration-improvement
Change-Id: I05ba531a18ba416e8b07c8a4165f5e1f7f5e44aa