Updated openstack/openstack Project: openstack/horizon  84aae88057b30460ea7616c253afc4726bf28eb4
Improving readability for Horizon top module structure

This patch improves the readability on how Horizon's application
level module `hz.app` is composited and how its dependency modules
are organized.  The dependency modules come from three sources:

1) Library modules - modules defined in third-party libraries,
   including angular's extensions
2) Horizon's built-in modules, including modules from `framework`
   components and modules from `openstack_dashboard` application
   core components.
3) Horizon's plug-in modules - modules from Horizon's extensions
   components via Horizon's plug-in mechanism.

Horizon's built-in modules are managed as a couple of trees, then only
the top modules of each tree should be listed in the Horizon's built-in
modules.

Change-Id: I989a0b5cc8b24f6b168629f4a8fa5e485dcf1a2d
Closes-Bug: #1472098