Updated openstack/openstack Project: openstack/cinder  24e7cbd0a0554d0c2a38ad110c0780e09d7bd4b1
Fix error message in Pure driver with correct text

The error message was specifically mentioning the PureISCSIDriver but
now there is a PureFCDriver which could call the error message. This
change gets the name of the driver and uses that to ensure
the error message refers to the correct driver.

Closes-Bug: #1469523
Change-Id: Ib724073ea352dadfbe7ddcf4ee4ceedd5d2c7c41