Updated openstack/openstack Project: openstack/swift  be757e7ae1884f79ad7f7d2b0eb2cfb144024b6c
Add CORS unit tests to base

In earlier versions of swift when a request was made with an
existing origin, but without any CORS settings in the container,
it was possible to get an unhandled exception due to a method call
on the "None" return of cors.get('allow_origin', '').

Unit tests have been added to assert that this problem cannot go
undetected again.

Change-Id: Ia74896dabe1cf5a307c551b15a43ab1fd789c213
Fixes: bug 1468782