Updated openstack/openstack Project: openstack/barbican  dcbff789f33564010a33ee8ece3972eaf671b381
Remove invalid skipping of tests

One of the tests in the test_validators suite was being skipped in
reference to a bug that was deemed invalid. So skipping it wasn't the
appropriate thing to do. Now that's not the case.

The other one was reffering to a feature that has already been
implemented, therefore we can stop skipping that test.

Change-Id: Ia068ff04ab1ea79b77727ad388720692c83df8cf